requirements.txt
pandas==2.8.1
numpy==1.24.2
sqlalchemy==2.0.7
snowflake-connector-python==3.3.1
snowflake-sqlalchemy==1.3.3
mysql-connector-python==8.0.32
requests==2.28.2
pyyaml==6.0
mage-ai==0.10.3
matplotlib==3.7.1
seaborn==0.12.2
jupyter==1.0.0
scikit-learn==1.2.2
tqdm==4.65.0
